add qwen2 & qwen3next & qwen3vl model testcase

+++ test/registered/ascend/embedding_models/test_ascend_embedding_models.py
@@ -16,12 +16,11 @@ register_npu_ci(
     disabled="embeddings are not all close",
 )
 
-
 MODELS = [
     ("/root/.cache/modelscope/hub/models/iic/gte_Qwen2-1.5B-instruct", 1, 1e-5),
     ("/root/.cache/modelscope/hub/models/Qwen/Qwen3-Embedding-8B", 1, 1e-5),
 ]
-TORCH_DTYPES = [torch.bfloat16]
+TORCH_DTYPES = [torch.float16]
 
 
 class TestEmbeddingModels(CustomTestCase):
